feat(withdraw): create withdraw feature

+++ database/migrations/2023_01_26_155238_create_finances_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
                 'status',
                 ['PENDING', 'SUCCESS', 'FAILED', 'ACCEPT', 'REJECT']
             )->index();
-            $table->unsignedInteger('balance');
+            $table->unsignedBigInteger('balance');
             $table->timestamps();
         });
     }
